+++ src/main/java/org/utn/edu/ar/model/exceptions/player/PlayerNotFoundException.java
@@ -5,7 +5,7 @@ package org.utn.edu.ar.model.exceptions.player;
  */
 public class PlayerNotFoundException extends Exception {
 
-    public PlayerNotFoundException(Integer id){
+    public PlayerNotFoundException(Long id){
         super("Player with id: " + id + " not found.");
     }
     public PlayerNotFoundException(String fbId){
