+++ src/main/java/org/utn/edu/ar/model/persistence/IPlayerStorage.java
@@ -1,7 +1,6 @@
 package org.utn.edu.ar.model.persistence;
 
 import org.utn.edu.ar.model.domain.Player;
-import org.utn.edu.ar.model.domain.Sport;
 import org.utn.edu.ar.model.exceptions.player.PlayerAlreadyExistsException;
 import org.utn.edu.ar.model.exceptions.player.PlayerNotFoundException;
 
@@ -14,17 +13,17 @@ public interface IPlayerStorage {
 
     public List<Player> getAll();
 
-    public Player getById(Integer id) throws PlayerNotFoundException;
+    public Player getById(Long id) throws PlayerNotFoundException;
 
-    public boolean exists(Integer id);
+    public boolean exists(Long id);
 
     public boolean exists(String fbId);
 
     public Player create(String fbId) throws PlayerAlreadyExistsException;
 
-    public void update(Integer id, String fbId) throws PlayerNotFoundException;
+    public void update(Long id, String fbId) throws PlayerNotFoundException;
 
-    public void remove(Integer id) throws PlayerNotFoundException;
+    public void remove(Long id) throws PlayerNotFoundException;
 
     public Player getByFacebookId(String id) throws PlayerNotFoundException;
 }
