+++ src/main/java/org/utn/edu/ar/model/persistence/memoryStorage/SportStorage.java
@@ -24,7 +24,7 @@ public class SportStorage implements ISportStorage {
     }
 
     @Override
-    public Sport getSportById(Integer id) {
+    public Sport getSportById(Long id) {
         Sport outputSport = null;
         for (Sport currentSport : sportList) {
             if (id.equals(currentSport.getId())) {
@@ -44,7 +44,7 @@ public class SportStorage implements ISportStorage {
     }
 
     @Override
-    public void updateSport(Integer sportId, String sportName) {
+    public void updateSport(Long sportId, String sportName) {
         removeSport(sportId);
         Sport currentSport = new Sport(sportId, sportName);
         sportList.add(currentSport);
@@ -52,14 +52,14 @@ public class SportStorage implements ISportStorage {
 
     @Override
     public Sport createSport(String sportName) {
-        int lastId = getLastId();
+        long lastId = getLastId();
         Sport currentSport = new Sport(lastId, sportName);
         sportList.add(currentSport);
         return currentSport;
     }
 
     @Override
-    public boolean exists(Integer sportId) {
+    public boolean exists(Long sportId) {
         boolean found = false;
         Iterator<Sport> itSport = sportList.iterator();
         while (itSport.hasNext() && !found) {
@@ -71,13 +71,13 @@ public class SportStorage implements ISportStorage {
         return found;
     }
 
-    public void removeSport(Integer sportId) {
+    public void removeSport(Long sportId) {
         Sport sportToRemove = new Sport(sportId);
         sportList.remove(sportToRemove);
     }
 
-    private int getLastId() {
-        int lastInt = 0;
+    private long getLastId() {
+        long lastInt = 0;
 
         for (Sport currentSport : sportList) {
             if (currentSport.getId() > lastInt) {
